Hold the EasyPowerPoint binding surface against the published one
The Windows CI run found 18 failing compatibility tests. Checking properly turned
up 44 differences from BalaReva.EasyPowerPoint.Activities 11.0.0, every one of
which would break a workflow binding to the published names:

- Four classes were never written at all: PowerPointObject, SlideObject,
  TextShape, ShapeFont and ImageSize. PowerPointScope.Body took an
  ActivityAction<string> rather than one over PowerPointObject, so the body got
  a file path where the published package hands over the presentation.
- Twelve position and size arguments were double where PowerPoint measures in
  points and the published package declared float.
- Eleven switches were InArgument<bool> where the published package declared a
  plain property, and eight more were bool where it declared the three-state
  TrueFalseNoneEnum. None means "leave the setting alone", which InArgument<bool>
  cannot express, so Print and StyleOption were overwriting settings a workflow
  never asked them to touch.
- Six outputs and inputs had the wrong shape: SlideExtractor saved a file rather
  than reporting a slide's text shapes, FindText reported matched text rather
  than the slides matched on, CommentsRead joined the comments into one string,
  ReadTables and RefreshData took one slide where the published package took an
  array, and ExtractTables returned a DataSet rather than a DataTable[].

Also adds audit/verify_binding_surface.py, which is how the rest were found. It
reads type, property and enum names back out of a built .nupkg and holds them
against the surface recorded off the gallery. Metadata only, so it runs on Linux,
and CI now runs it on the Linux job right after packing rather than leaving this
to the Windows test run. It resolves inherited properties up the base chain: a
property moved onto a shared base is still the same property to a workflow. All
seven packages pass it.

PowerPointObject.PptPersentation is a Microsoft.Office.Interop.PowerPoint
.Presentation in the published package, so the package now references that
interop assembly - for the type identity alone; nothing is called through it, and
PowerPointService stays late-bound because Microsoft.Office.Core still has no
usable distribution. It is the same repackage family already relied on for Word
and Outlook, unlike the Office 2007-era Microsoft.Office.Core package this
repository's own audit stance rules out.

The Excel bridging activities now refuse in the activity rather than in the
service, so the refusal holds for any implementation and the test can see it;
the earlier test passed vacuously against a stand-in that recorded the call
instead of throwing.

Two smaller fixes: RequireProperties now names the property and both types in its
failure message, which would have made the CI run readable, and two generated
files had ' escapes where an apostrophe belonged.

+++ src/BalaReva.EasyPowerPoint/Scope/Chart/ChartShapeCountAndFriends.cs
@@ -76,25 +76,25 @@ public sealed class ChartFormat : BaseNativeChild
     [Category("Input")]
     [DisplayName("Chart Left")]
     [Description("Distance from the left edge of the slide, in points. Zero leaves it alone.")]
-    public InArgument<double> ChartLeft { get; set; } = null!;
+    public InArgument<float> ChartLeft { get; set; } = null!;
 
     /// <summary>Distance from the top edge of the slide, in points.</summary>
     [Category("Input")]
     [DisplayName("Chart Top")]
     [Description("Distance from the top edge of the slide, in points. Zero leaves it alone.")]
-    public InArgument<double> ChartTop { get; set; } = null!;
+    public InArgument<float> ChartTop { get; set; } = null!;
 
     /// <summary>Width in points.</summary>
     [Category("Input")]
     [DisplayName("Chart Width")]
     [Description("Width in points. Zero leaves it alone.")]
-    public InArgument<double> ChartWidth { get; set; } = null!;
+    public InArgument<float> ChartWidth { get; set; } = null!;
 
     /// <summary>Height in points.</summary>
     [Category("Input")]
     [DisplayName("Chart Height")]
     [Description("Height in points. Zero leaves it alone.")]
-    public InArgument<double> ChartHeight { get; set; } = null!;
+    public InArgument<float> ChartHeight { get; set; } = null!;
 
     /// <inheritdoc />
     protected override void ExecuteWork(CodeActivityContext context, IPowerPointPresentation presentation) =>
